+++ bin/.gitignore
@@ -1,1 +1,3 @@
+/module-info.class
 /net/
+/plugin.yml
