+++ src/module-info.java
@@ -1,0 +1,14 @@
+module pgfCore {
+	
+	exports net.pgfmc.core.requestAPI;
+	exports net.pgfmc.core.inventoryAPI;
+	exports net.pgfmc.core.configify;
+	exports net.pgfmc.core.playerdataAPI;
+	exports net.pgfmc.core.misc;
+	
+	
+	requires transitive org.bukkit;
+	
+	
+	
+}
