+++ src/net/pgfmc/core/misc/CoreMain.java
@@ -1,4 +1,4 @@
-package net.pgfmc.core;
+package net.pgfmc.core.misc;
 
 import java.io.File;
 import java.util.HashMap;
@@ -19,12 +19,12 @@ import org.bukkit.scoreboard.Scoreboard;
 import net.pgfmc.core.cmd.Blocked;
 import net.pgfmc.core.cmd.Goto;
 import net.pgfmc.core.cmd.admin.Broadcast;
-import net.pgfmc.core.cmd.admin.Skull;
+import net.pgfmc.core.cmd.admin.SkullCommand;
 import net.pgfmc.core.cmd.admin.Tagging;
 import net.pgfmc.core.cmd.donator.Nick;
 import net.pgfmc.core.configify.ReloadConfigify;
 import net.pgfmc.core.inventoryAPI.InventoryPressEvent;
-import net.pgfmc.core.permissions.PermissionsManager;
+import net.pgfmc.core.playerdataAPI.PermissionsManager;
 import net.pgfmc.core.playerdataAPI.PlayerDataManager;
 
 /**
@@ -158,7 +158,7 @@ public class CoreMain extends JavaPlugin implements Listener {
 		
 		getCommand("nick").setExecutor(new Nick());
 		
-		getCommand("skull").setExecutor(new Skull());
+		getCommand("skull").setExecutor(new SkullCommand());
 		
 		getCommand("pgf").setExecutor(new ReloadConfigify());
 		getCommand("tag").setExecutor(new Tagging());
