+++ src/net/pgfmc/core/requestAPI/Requester.java
@@ -1,6 +1,5 @@
 package net.pgfmc.core.requestAPI;
 
-import java.net.http.WebSocket.Listener;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
@@ -11,9 +10,10 @@ import java.util.stream.Collectors;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerQuitEvent;
 
-import net.pgfmc.core.CoreMain;
+import net.pgfmc.core.misc.CoreMain;
 import net.pgfmc.core.playerdataAPI.PlayerData;
 
 /**
